Read included and excluded way tags from args
Add two new command line flags: `--includeWays` and `--excludeWays`. Include can
be used as a "if any of these match, include the tag" whereas exclude can be
used as a "if any of these match, do not include the tag". Shorthands are -i and
-e.

Usage:
  osmparser -f map.osm -i highway railway=abandoned -e access=no access=private
  osmparser -f map.osm -i railway -e railway=abandoned

Commit also removes --verbose and adds --quiet. Proper logging should be added
next.